agents/vision-agent/requirements.txt
a2a-sdk[http-server]
google-genai
uvicorn
Pillow
pydantic
python-dotenv
